% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/an_colors.R
\docType{data}
\name{an_colors}
\alias{an_colors}
\title{A list of colors I want regular access to.}
\format{
An object of class \code{list} of length 3.
}
\usage{
an_colors
}
\description{
A list of colors I want regular access to.
}
\examples{
an_colors$blue
}
\keyword{datasets}
