% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/an_save.R
\name{an_save}
\alias{an_save}
\title{Wrapper for ggsave with some defaults.}
\usage{
an_save(filename, shape = "square")
}
\arguments{
\item{filename}{filename to save the file to}

\item{shape}{options: "square", "hrect", "vrect"; default = "square"}
}
\value{
nothing; saves a file at the specified filepath
}
\description{
Wrapper for ggsave with some defaults.
}
\examples{
\dontrun{
an_save("test.png")
an_save("test.png", shape = "hrect")
}
}
