% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anRpackage.R
\docType{package}
\name{anRpackage}
\alias{anRpackage}
\alias{anRpackage-package}
\title{anRpackage: Testing to see if I can get Rcpp Armadillo to work with other basic 
r functions for a package.}
\section{anRpackage functions}{

writing some functions
}

