% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main2.R
\name{as.matrix.ecological_inference_model}
\alias{as.matrix.ecological_inference_model}
\title{Convert EMModel Object to a Probability Matrix}
\usage{
\method{as.matrix}{ecological_inference_model}(object, ...)
}
\value{
A matrix containing the estimated probabilities.
}
\description{
Extracts the probability matrix from the model, making it able to manipulate it as
a matrix
}
