% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxc-package.R
\name{as.matrix.rxc}
\alias{as.matrix.rxc}
\title{Extract Probability Matrix from rxc Object}
\usage{
\method{as.matrix}{rxc}(object)
}
\arguments{
\item{object}{An rxc object.}
}
\value{
A matrix containing the estimated probabilities.
}
\description{
Extracts the probability matrix from an rxc object.
}
\examples{
model <- rxc$new(X = matrix(1:15, 5, 3), W = matrix(1:50, 5, 10))
as.matrix(model)
}
