% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main2.R
\name{ballots.ecological_inference_model}
\alias{ballots.ecological_inference_model}
\title{Returns the current amount of ballots boxes.}
\usage{
\method{candidates}{ecological_inference_model}(object)
}
\arguments{
\item{object}{An `ecological_inference_model` object.}
}
\value{
The amount of ballots.
}
\description{
Given a initialized object, it returns the amount of ballot boxes it has. It's equivalent
of running nrow(object$W) or nrow(object$X).
}
