% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxc-package.R
\name{ballots.rxc}
\alias{ballots.rxc}
\title{Get Number of Ballot Boxes}
\usage{
ballots.rxc(object)
}
\arguments{
\item{object}{An rxc object.}
}
\value{
An integer representing the number of ballots.
}
\description{
Returns the number of ballot boxes in an rxc object.
}
\examples{
model <- rxc$new(X = matrix(1:15, 5, 3), W = matrix(1:50, 5, 10))
ballots(model)
}
