% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main2.R
\name{candidates.ecological_inference_model}
\alias{candidates.ecological_inference_model}
\title{Returns the current amount of candidates.}
\usage{
\method{candidates}{ecological_inference_model}(object)
}
\arguments{
\item{object}{An `ecological_inference_model` object.}
}
\value{
The amount of candidates.
}
\description{
Given a initialized object, it returns the amount of candidates that it has. It's equivalent
of running ncol(object$X).
}
