% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxc-package.R
\name{candidates.rxc}
\alias{candidates.rxc}
\title{Get Number of Candidates}
\usage{
candidates.rxc(object)
}
\arguments{
\item{object}{An rxc object.}
}
\value{
An integer representing the number of candidates.
}
\description{
Returns the current number of candidates in an rxc object.
}
\examples{
model <- rxc$new(X = matrix(1:15, 5, 3), W = matrix(1:50, 5, 10))
candidates(model)
}
