% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designfunctions.R
\name{create_designmatrix}
\alias{create_designmatrix}
\title{Generates a matrix to describe the relationship between the domains}
\usage{
create_designmatrix(X, Y, DENS, rowname = "X", colname = "Y")
}
\arguments{
\item{X}{Number of rows}

\item{Y}{Number of columns}

\item{DENS}{if = 2: diagonal matrix, if = -1: random density between the desired boundaries}

\item{rowname}{}

\item{colname}{}
}
\value{
Matrix which contains 1 and 0 entries
}
\description{
Generates a matrix to describe the relationship between the domains
}
