% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main2.R
\docType{class}
\name{EMModel}
\alias{EMModel}
\title{EMModel: An R6 Class for Running an Expectation-Maximization Algorithm}
\description{
EMModel: An R6 Class for Running an Expectation-Maximization Algorithm

EMModel: An R6 Class for Running an Expectation-Maximization Algorithm
}
\details{
This class implements an EM algorithm using different methods for approximating the E-step such as "Multinomial",
"Hit and Run", "MVN CDF", "MVN PDF", and "Exact".
}
\note{
The method attribute wont be updated, since the main computation haven't been called and it's possible (but weird) that the
user may want to run another method even if the precomputation was made.

For having the finalizer on the private access, R6 must be >= 2.4.0.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{X}}{A (b x c) matrix with the observed results of the candidate votes (c) on a given
ballot box (b). Provided manually or loaded from JSON.}

\item{\code{W}}{A (b x g) matrix with the observed results of the demographical group votes (g) on
a given ballot box (b). Provided manually or loaded from JSON.}

\item{\code{method}}{A string indicating the EM method. One of: "Multinomial", "Hit and Run", "MVN CDF",
"MVN PDF", "Exact".}

\item{\code{probability}}{A (g x c) matrix that would store the final estimated probabilities of having
a given group (g) voting for a candidate (c).}

\item{\code{logLikelihood}}{A numeric vector that will store the log-likelihood values among the total
iterations of the Expected Maximization algorithm.}

\item{\code{total_iterations}}{An integer indicating the total iterations that the Expected Maximization
algorithm did.}

\item{\code{total_time}}{The time that the EM algorithm took.
Constructor for the EMModel class.}

\item{\code{been_called}}{Boolean that determines if the object has been called before. It's mainly used for triggering
the C cleanup and updating its global variables}

\item{\code{been_precomputed_hr}}{Boolean that determines if the object has been precomputed for the Hit and Run method.}

\item{\code{been_precomputed_hr}}{Boolean that determines if the object has been precomputed for the Exact method.}

\item{\code{empty_candidates}}{Integer vector with the index of candidates that didn't receive any vote.
It's used for handling border cases and optimizing.}

\item{\code{been_computed}}{Boolean that determines if the EM-algorithm have been computed.}

\item{\code{hr_samples}}{The samples used on the computation of the EM-algorithm under the Hit and Run method.}

\item{\code{hr_step_size}}{The step size used on the computation of the EM-algorithm under the Hit and Run method.}

\item{\code{mvn_method}}{Save the method used to calculate the multivariate normal cdf.}

\item{\code{mvn_error}}{Saves the error used for estimating the Montecarlo simulation of the multivariate normal cdf.}

\item{\code{mvn_iterations}}{Saves the maximum iterations made to estimate the Montecarlo simulation of the
multivariate normal cdf.
Destructor to clean allocated memory

When the object gets removed or garbage collected, it cleans the allocated memory on C that could had been
reused for the object instance. It's used for avoiding memory leaks due to the C behavior.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{samples}}{Active variable to show the Hit and Run samples if and only if self$method is "Hit and Run"}

\item{\code{step_size}}{Active variable to show the Hit and Run step size if and only if self$method is "Hit and Run"}

\item{\code{multivariate_method}}{Active variable to show the method used to estimate the Multivariate
Normal CDF, if and only if self$method is "MVN CDF"}

\item{\code{multivariate_error}}{Active variable to show the error threshold for the Montecarlo simulation of the
Multivariate Normal CDF}

\item{\code{multivariate_error}}{Active variable to show the error threshold for the Montecarlo simulation of the
Multivariate Normal CDF}

\item{\code{been_called}}{Boolean that determines if the object has been called before. It's mainly used for triggering
the C cleanup and updating its global variables}

\item{\code{been_precomputed_hr}}{Boolean that determines if the object has been precomputed for the Hit and Run method.}

\item{\code{been_precomputed_hr}}{Boolean that determines if the object has been precomputed for the Exact method.}

\item{\code{empty_candidates}}{Integer vector with the index of candidates that didn't receive any vote.
It's used for handling border cases and optimizing.}

\item{\code{been_computed}}{Boolean that determines if the EM-algorithm have been computed.}

\item{\code{hr_samples}}{The samples used on the computation of the EM-algorithm under the Hit and Run method.}

\item{\code{hr_step_size}}{The step size used on the computation of the EM-algorithm under the Hit and Run method.}

\item{\code{mvn_method}}{Save the method used to calculate the multivariate normal cdf.}

\item{\code{mvn_error}}{Saves the error used for estimating the Montecarlo simulation of the multivariate normal cdf.}

\item{\code{mvn_iterations}}{Saves the maximum iterations made to estimate the Montecarlo simulation of the
multivariate normal cdf.
Destructor to clean allocated memory

When the object gets removed or garbage collected, it cleans the allocated memory on C that could had been
reused for the object instance. It's used for avoiding memory leaks due to the C behavior.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ecological_inference_model-new}{\code{EMModel$new()}}
\item \href{#method-ecological_inference_model-precompute}{\code{EMModel$precompute()}}
\item \href{#method-ecological_inference_model-compute}{\code{EMModel$compute()}}
\item \href{#method-ecological_inference_model-print}{\code{EMModel$print()}}
\item \href{#method-ecological_inference_model-summary}{\code{EMModel$summary()}}
\item \href{#method-ecological_inference_model-save_results}{\code{EMModel$save_results()}}
\item \href{#method-ecological_inference_model-clone}{\code{EMModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ecological_inference_model-new"></a>}}
\if{latex}{\out{\hypertarget{method-ecological_inference_model-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMModel$new(X = NULL, W = NULL, jsonPath = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ecological_inference_model-precompute"></a>}}
\if{latex}{\out{\hypertarget{method-ecological_inference_model-precompute}{}}}
\subsection{Method \code{precompute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMModel$precompute(method, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ecological_inference_model-compute"></a>}}
\if{latex}{\out{\hypertarget{method-ecological_inference_model-compute}{}}}
\subsection{Method \code{compute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMModel$compute(
  main_method = "Multinomial",
  probability_method = "Group proportional",
  iterations = 1000,
  stopping_threshold = 0.001,
  verbose = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ecological_inference_model-print"></a>}}
\if{latex}{\out{\hypertarget{method-ecological_inference_model-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMModel$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ecological_inference_model-summary"></a>}}
\if{latex}{\out{\hypertarget{method-ecological_inference_model-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMModel$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ecological_inference_model-save_results"></a>}}
\if{latex}{\out{\hypertarget{method-ecological_inference_model-save_results}{}}}
\subsection{Method \code{save_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMModel$save_results(filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ecological_inference_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ecological_inference_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EMModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
