% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Title euclidean}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{}

\item{y}{}
}
\value{
r_2
}
\description{
Title euclidean
}
\examples{
euclidean(123612,13892347912)
}
