% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDemand.R
\name{gen_Demand}
\alias{gen_Demand}
\title{generate Demand}
\usage{
gen_Demand(NUMB_C, TQ, Q_VAR)
}
\arguments{
\item{NUMB_C}{Length of the series of numbers}

\item{TQ}{Sum of the series of numbers}

\item{Q_VAR}{??}
}
\value{
Vektor which contains the randomly generated values
}
\description{
Generates a random series of numbers, from which the sum is normalized to TQ and returned
}
