% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_RCC.R
\name{gen_RCC}
\alias{gen_RCC}
\title{Generate resource cost vektor}
\usage{
gen_RCC(RC_VAR, TC, NUMB_RC)
}
\arguments{
\item{RC_VAR}{??}

\item{TC}{Sum of the series of numbers}

\item{NUMB_RC}{Length of the series of numbers}
}
\value{
Vektor which contains the randomly generated values
}
\description{
Generates a random series of numbers, from which the sum is normalized to TC and returned
}
