% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_RES_CONS_PAT.R
\name{gen_RES_CONS_PAT}
\alias{gen_RES_CONS_PAT}
\title{Generates a matrix to define the relationship between CM and PV}
\usage{
gen_RES_CONS_PAT(EAD, NUMB_CM, NUMB_PV, DENS_CMPV)
}
\arguments{
\item{EAD}{}

\item{NUMB_CM}{}

\item{NUMB_PV}{}

\item{DENS_CMPV}{}
}
\value{
EAD??
}
\description{
Random matrix with entries greater than 1 is generated somehow
}
