% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxc-package.R
\name{groups.rxc}
\alias{groups.rxc}
\title{Get Number of Groups}
\usage{
groups.rxc(object)
}
\arguments{
\item{object}{An rxc object.}
}
\value{
An integer representing the number of groups.
}
\description{
Returns the current number of groups in an rxc object.
}
\examples{
model <- rxc$new(X = matrix(1:15, 5, 3), W = matrix(1:50, 5, 10))
groups(model)
}
