% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood.R
\name{likelihood}
\alias{likelihood}
\title{Calculate the Likelihood of Rasch data}
\usage{
likelihood(raschObj, theta)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{theta}{A numeric representation of a person's ability}
}
\value{
The calculated likelihood from Probability function
}
\description{
Returns the likelihood the observed value in the Probability function occurs
}
\note{
This is a help session file
}
\author{
Dominique Lockett
}
