% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someCode.R
\name{multiply_numbers}
\alias{multiply_numbers}
\title{Multiply two numbers}
\usage{
multiply_numbers(x, y)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The product of x and y.
}
\description{
This function multiplies two numbers.
}
\examples{
multiply_numbers(2, 3)

}
