% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main2.R
\name{predict.ecological_inference_model}
\alias{predict.ecological_inference_model}
\title{Predict Probability using an EM Model}
\usage{
\method{predict}{ecological_inference_model}(object, ...)
}
\value{
A matrix of estimated probabilities.
}
\description{
This method is a wrapper around `compute()`, allowing prediction of probabilities
based on specified parameters. The results are stored in `object$probability`. Refer
to object$compute method for more information about its function and parameters. However
it will esentially trigger the EM algorithm.
}
