% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.R
\docType{methods}
\name{print,ANY-method}
\alias{print,ANY-method}
\title{Print}
\usage{
\S4method{print}{ANY}(raschObj)
}
\arguments{
\item{name}{The name of the person being analyzed}

\item{EAP}{Expected a Priori value}
}
\value{
A single numeric value that is is the negative LL
}
\description{
Returns the height of the LL for a  given value of lambda
}
\note{
This is a help session file
}
\author{
Dominique Lockett
}
