% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{prior}
\alias{prior}
\title{Prior}
\usage{
prior(theta)
}
\arguments{
\item{theta}{A numeric object representing a person's ability}
}
\value{
The height of the specified normal curve evaluated at theta
}
\description{
Returns the height of a normal curve evaluated at theta
}
\note{
This is a help session file
}
\author{
Dominique Lockett
}
