% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main2.R
\name{sum.ecological_inference_model}
\alias{sum.ecological_inference_model}
\title{Returns the total amount of voters in the system.}
\usage{
\method{candidates}{ecological_inference_model}(object)
}
\arguments{
\item{object}{An `ecological_inference_model` object.}
}
\value{
The amount of voters.
}
\description{
Given a initialized object, it returns the total amount of voters. It's equivalent
of running sum(object$W) or sum(object$X)
}
