% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxc-package.R
\name{sum.rxc}
\alias{sum.rxc}
\title{Get Total Number of Voters}
\usage{
\method{sum}{rxc}(object)
}
\arguments{
\item{object}{An rxc object.}
}
\value{
An integer representing the total number of voters.
}
\description{
Returns the total number of voters in the system from an rxc object.
}
\examples{
model <- rxc$new(X = matrix(1:15, 5, 3), W = matrix(1:50, 5, 10))
sum(model)
}
