% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main2.R
\name{summary.ecological_inference_model}
\alias{summary.ecological_inference_model}
\title{Summarize the object main attributes.}
\usage{
\method{summary}{ecological_inference_model}(object, ...)
}
\description{
This method is an S3 styled wrapper around `summary()`, allowing to get, as a list, the
main details of the object. Additional parameters are shown depending on the state of
the object (computed/uncomputed).
}
