% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxc-package.R
\name{summary.rxc}
\alias{summary.rxc}
\title{Summarize rxc Object}
\usage{
\method{summary}{rxc}(object)
}
\arguments{
\item{object}{An rxc object.}
}
\value{
A list with the principal attributes.
}
\description{
Returns a list with the principal attributes of an rxc object.
}
\examples{
model <- rxc$new(X = matrix(1:15, 5, 3), W = matrix(1:10, 2, 5))
summary(model)
}
