% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_an.R
\name{theme_an}
\alias{theme_an}
\title{Personal theme.}
\usage{
theme_an(axes = TRUE)
}
\arguments{
\item{axes}{whether to include axes or not}
}
\value{
ggplot2 theme object
}
\description{
Personal theme.
}
\examples{
theme_an()
}
