% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main2.R
\name{update.ecological_inference_model}
\alias{update.ecological_inference_model}
\title{Update an existing RxG model with a new EM algorithm computation}
\usage{
\method{update}{ecological_inference_model}(object, ...)
}
\arguments{
\item{object}{An `ecological_inference_model` object.}

\item{...}{New parameters to pass to `compute()`.}
}
\value{
The updated object.
}
\description{
This function updates an object with a new Expected Maximization computation
with other parameters.
}
