% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxc-package.R
\name{update.rxc}
\alias{update.rxc}
\title{Update an rxc Object}
\usage{
\method{update}{rxc}(object, ...)
}
\arguments{
\item{object}{An rxc object.}

\item{...}{New parameters to pass to \code{compute()}.}
}
\value{
The updated rxc object.
}
\description{
Updates an existing rxc model with a new EM algorithm computation.
}
\examples{
model <- rxc$new(X = matrix(1:15, 5, 3), W = matrix(1:50, 5, 10))
update(model, "MVN PDF")
}
