\name{var1}
\alias{var1}
\docType{data}
\title{
var1: example variable one. A sample from a normal distribution
}
\description{
example variable one. A sample from a normal distribution
}
\usage{data(var1)}
\format{
  The format is:
 num [1:100] 6.38 10.83 13.25 2.96 11.29 ...
}
\details{
none}
\source{
simulated data
}
\references{
none so far}
\examples{
data(var1)
## maybe str(var1) ; plot(var1) ...
}
\keyword{datasets}
