\name{var2}
\alias{var2}
\docType{data}
\title{
var2: example variable one. A sample from a normal distribution
}
\description{
example variable one. A sample from a normal distribution
}
\usage{data(var2)}
\format{
  The format is:
 num [1:100] 17.3 18.3 15.1 18.3 11.2 ...
}
\details{
None
}
\source{
simulated data
}
\references{
None
}
\examples{
data(var2)
## maybe str(var2) ; plot(var2) ...
}
\keyword{datasets}
