\name{Fac}
\alias{ttsis-package}
\alias{ttsis}
\docType{package}
\title{
\packageTitle{anRpackage}
}
\description{
\packageDescription{a package to caculate factorials}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{a package to caculate factorials}
\packageIndices{Fac}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Marui Raojialu Heyaqi}

Maintainer: \packageMaintainer{Marui}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
Fac(3)
}
