\name{anRpackage-package}
\alias{anRpackage-package}
\alias{anRpackage}
\docType{package}
\title{
\packageTitle{anRpackage}
}
\description{
\packageDescription{anRpackage}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{anRpackage}

Maintainer: \packageMaintainer{anRpackage}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = anRpackage)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{anRpackage}}
%\section{Documentation Index}{\packageIndices{anRpackage}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
